🚚 Move utils function into their own files
