+++ src/utils/padNumber.js
@@ -1,0 +1,3 @@
+export default function padNumber(x) {
+  return (x.toString().length < 2 ? `0${x}` : x).toString();
+}
