+++ src/utils/uncapitalize.js
@@ -1,0 +1,3 @@
+export default function uncapitalize(string) {
+  return string[0].toLowerCase() + string.slice(1);
+}
